race-service/requirements.txt
Django>=2.2,<2.3
djangorestframework>=3.9.2,<3.10.0

drf-extensions>=0.5.0,<0.6.0
psycopg2>=2.7.5,<2.8.0
requests>=2.21.0,<2.30.0

django-extensions
flake8>=3.7.7,<3.8.0
pytest>=4.4.1,<4.5.0
pytest-django>=3.4.8,<3.5.0
factory-boy>=2.12.0,<2.13.0